Update Generate.ipynb
Now more that 4 words can be generated, and if the 3 last words are not found in quadgrams, it looks for the last 2 words in trigrams, if not found it looks for the last word in bigrams, if not found it randomly generates a word from unigrams.  It is still perdictable because it uses the most common occurance.  I will try to make it pick randomly from more options