Update Generate.ipynb again . . .
I fixed some cutting and pasting that I had done that referred to an old variable that I didn't even initaite in the code.  I discovered this by running it in a new jupyter notebook.  That explains why it was only using 4 or less occurances.  Now it takes a little longer to run and not sure if the output makes more sence, but it should be even less predicatable now.